+++ IspDataDb/dbo/Tables/Deduction.sql
@@ -4,7 +4,7 @@
     [DeductionName]    VARCHAR (50)   NULL,
     [DeductionNameAra] NVARCHAR (50)  NULL,
     [DefaultFrequency] CHAR (1)       NULL,
-    [AccountIdNo]      INT            NULL,
+    [AccountIdNo]      SMALLINT       NULL,
     [DeductionType]    CHAR (1)       NULL,
     [DeductionPlace]   CHAR (1)       NULL,
     [ComputationType]  CHAR (1)       NULL,
@@ -13,6 +13,8 @@
     [DateTimeStamp]    ROWVERSION     NULL,
     CONSTRAINT [PK_Deduction] PRIMARY KEY CLUSTERED ([IdNo] ASC)
 );
+
+
