+++ IspDataDb/dbo/Tables/EmployeeLoanJournal.sql
@@ -1,11 +1,11 @@
 ﻿CREATE TABLE [dbo].[EmployeeLoanJournal] (
     [IdNo]            INT            IDENTITY (1, 1) NOT NULL,
-    [EmployeeIdNo]    INT            NOT NULL,
+    [EmployeeIdNo]    SMALLINT       NOT NULL,
     [TransactionDate] DATE           NULL,
     [ReferenceNo]     VARCHAR (15)   COLLATE SQL_Latin1_General_CP1_CI_AS NULL,
     [TransactionType] CHAR (1)       COLLATE SQL_Latin1_General_CP1_CI_AS NULL,
     [Amount]          MONEY          NOT NULL,
-    [AccountIdNo]     INT            NOT NULL,
+    [AccountIdNo]     SMALLINT       NOT NULL,
     [Notes]           NVARCHAR (255) COLLATE SQL_Latin1_General_CP1_CI_AS NOT NULL,
     [Posted]          BIT            NULL,
     [Cancelled]       BIT            NULL,
@@ -13,4 +13,6 @@
     [DateTimeStamp]   ROWVERSION     NOT NULL,
     CONSTRAINT [PK_EmployeeLoanIdNo] PRIMARY KEY CLUSTERED ([IdNo] ASC)
 );
+
+
