+++ IspDataDb/dbo/Tables/PayInformation.sql
@@ -1,8 +1,10 @@
 ﻿CREATE TABLE [dbo].[PayInformation] (
     [IdNo]             INT      NULL,
-    [EmployeeIdNo]     INT      NULL,
+    [EmployeeIdNo]     SMALLINT NULL,
     [FrequencyOfPay]   CHAR (1) NULL,
     [SalariedOrHourly] CHAR (1) NULL,
     [Rate]             MONEY    NULL
 );
+
+
