+++ IspDataDb/dbo/Tables/PettyCashJournal.sql
@@ -3,7 +3,7 @@
     [TransactionDate]     DATE           NOT NULL,
     [ReferenceNo]         VARCHAR (15)   COLLATE SQL_Latin1_General_CP1_CI_AS NULL,
     [Amount]              MONEY          NOT NULL,
-    [AccountIdNo]         INT            NOT NULL,
+    [AccountIdNo]         SMALLINT       NOT NULL,
     [PaymentType]         CHAR (1)       COLLATE SQL_Latin1_General_CP1_CI_AS NULL,
     [PayeeIdNo]           INT            NULL,
     [PayeeName]           NVARCHAR (50)  COLLATE SQL_Latin1_General_CP1_CI_AS NULL,
@@ -11,7 +11,7 @@
     [CheckDate]           DATE           NULL,
     [ORNumber]            VARCHAR (15)   COLLATE SQL_Latin1_General_CP1_CI_AS NULL,
     [DiscountTaken]       MONEY          NULL,
-    [DiscountAccountIdNo] INT            NULL,
+    [DiscountAccountIdNo] SMALLINT       NULL,
     [Applied]             MONEY          NULL,
     [UnApplied]           MONEY          NULL,
     [VatNumber]           VARCHAR (15)   COLLATE SQL_Latin1_General_CP1_CI_AS NULL,
@@ -23,4 +23,6 @@
     [DateTimeStamp]       ROWVERSION     NULL,
     CONSTRAINT [PK_PettyCashJournal1] PRIMARY KEY CLUSTERED ([IdNo] ASC)
 );
+
+
