+++ IspDataDb/dbo/Tables/TableName.sql
@@ -1,10 +1,12 @@
 ﻿CREATE TABLE [dbo].[TableName] (
-    [IdNo]              BIGINT         IDENTITY (1, 1) NOT NULL,
+    [IdNo]              SMALLINT       IDENTITY (1, 1) NOT NULL,
     [DatabaseTableIdNo] SMALLINT       NULL,
-    [TableIdNo]         INT            NULL,
+    [TableIdNo]         SMALLINT       NULL,
     [Name]              NVARCHAR (100) NULL,
     CONSTRAINT [PK_TableName] PRIMARY KEY CLUSTERED ([IdNo] ASC)
 );
+
+
 
 
 GO
